Added error handling, made stuff pretier and added specific commands in package.json
Added error handling for when the specified runtime errors/isn't installed, changed if-else blocks to switch-case, added commands to run the client on bun, to run the client on node, to run the server on bun and to run the server on node, made default runtime for the server bun, added documentation about that to the readme. Added an entry to the config 'inputFile' which defines where the code input of the client is stored, also added documentation about that entry to the readme.